fix: refactore code
