+++ src/commands/commands.js
@@ -3,9 +3,8 @@ import readFile from "../fileOperations/readFile.js";
 import addEmptyFile from "../fileOperations/addEmptyFile.js";
 import renameFile from "../fileOperations/rename.js";
 import copyFile from "../fileOperations/copyFile.js";
-import { exitHandler } from "../utils/utils.js";
 import deleteFile from "../fileOperations/deleteFile.js";
-import { getOSData } from "../os/os-functions.js";
+import getOSData from "../os/os-functions.js";
 import calculateHash from "../hash/calculateHash.js";
 import compressFile from "../compress/compressFile.js";
 import decompressFile from "../compress/decompressFile.js";
