+++ src/userStore/user.js
@@ -1,0 +1,11 @@
+let user = '';
+
+const getUser = () => {
+    return user;
+};
+
+const setUser = (userName) => {
+    user = userName;
+};
+
+export { setUser, getUser };
