+++ src/utils/commandListener.js
@@ -1,30 +1,8 @@
 import readline from 'readline';
 import { commands } from '../commands/commands.js';
 import invalidCommandHandler from './invalidCommandHandler.js';
-
-let user = '';
-
-const initApp = () => {
-    const args = process.argv.slice(2);
-    const userName = args.join('').split('=')[1] || 'My dear user';
-    user = userName;
-    const welcomeWording = `Welcome to the File Manager, ${userName}!`;
-    console.log(welcomeWording);
-    printWorkingDirectory();
-}
-
-const printWorkingDirectory = () => {
-    const fileName = process.cwd();
-    const wording = `You are currently in ${fileName}`;
-    console.log(wording);
-};
-
-const exitHandler = () => {
-    const userName = user;
-    const exitWording = `Thank you for using File Manager, ${userName}, goodbye!`;
-    console.log(exitWording);
-    process.exit(0);
-};
+import exitHandler from './exitHandler.js';
+import printWorkingDirectory from './printWorkingDirectory.js';
 
 const commandListener = async () => {
     const readLine = readline.createInterface({
@@ -33,11 +11,21 @@ const commandListener = async () => {
       });
 
       const runCommand = async (line) => {
+        if (line === '.exit') {
+            exitHandler();
+        }
         const [first, ...rest] = line.split(' ');
         const commandFromInput = first;
         const command = commands.find((item) => Object.keys(item)[0] === commandFromInput);
         if (command) {
-            const args = command.defaultArgs ? [...rest, {defaultArgs: command.defaultArgs}] : rest;
+            let formattedArgs = rest
+            .join(' ')
+            .split('\'')
+            .filter((arg) => (arg !== '') && (arg !== ' '))
+            .map((arg) => {
+                return arg.replace(/\//g, "\\");
+            });
+            const args = command.defaultArgs ? [...formattedArgs, {defaultArgs: command.defaultArgs}] : formattedArgs;
             await command[commandFromInput](args); 
         } else {
             invalidCommandHandler();
@@ -50,13 +38,8 @@ const commandListener = async () => {
       }); 
 
     readLine.on('close', async () => {
-        console.log('close');
         exitHandler();
     });
 };
 
-export {    
-    exitHandler,
-    commandListener,
-    initApp,
-};
+export default commandListener;
